Fixing deletion bug in sound system
Because multiple sound subclients share a common source of data and
work asynchrounously and with minimal supervision, deleting streaming
sources is non-trivial.
Because of that, simon would crash when multiple resources used a
streaming input - usually the webservice tts provider.
This patch fixes the problem by introducing a QSharedPointer API.