Fixing volume calibration on start screen, bugfix with scenario display
Volume calibration is now not directly started from the Qt event handler
This fixes a memory corruption when opening the audio device failed
because of the dialog that pops up.
Fixed the problem that changing the scenario selection would add
to the old selection in the ui in the welcome page instead of replacing
it.