Made sam compatible with the new infrastructure
I had to break the sbm abstraction and work much closer with the
underlying technology (HTK / Julius for now) as SAM needs to operate
with the "real" model input files to give the user greater control.

This results in pretty bad code and usability but also a much more
powerful tool for powerusers.