Make the new welcome-page a bit prettier
This patch puts icons for each of the 4 categories on the welcome-screen in the
background of each (equal to what amarok does and kdevelop will do in it's new
welcome-screen). This helps a bit to seperate the categories better, but more
importantly it simply (partially) fetches back the wow-factor the old welcome
screen had.

For the background-images custom pixmaps are used that are based on oxygen-icons
and were modified like this (using gimp):
- (not bg-scenarios) rotate by -10deg (bg-acoustic-model and bg-training)
   or 45deg (bg-recognition)
- (only bg-acoustic-model) scale the layer to fit the image-size after rotating
- autotrim the image
- cut off 10px from the top and the left side.
- add a layer mask
- mark everything except 1px top and left and invert the selection
- colour the top-left-border that should now be marked black
   (=> make it transparent)
- apply a gaussian blur (radius: 2px) to the layer mask
- apply the layer mask
- set the opacity of the layer to 40%
(the bg-acoustic-model is based on the original svg, as there is no 128px
 version of it)