Fixing synchronization fork
When neither server nor client have an active model, the synchronization
may "fork" into two different, concurrent sub-synchronizations due to
a faulty handling of the case on the client. The server sends two
messages to the client: One to let him know that no model is available
and a second one to request the base model date.
The client - up until this revision - responded to the first (error-)
message with the next synchronization step, thefore creating another
synchronization "branch".