Bugfixing the calculation of deletion errors
Deletion errors are internally represented as notes that carry the label
of the word that is not present in the result and have the "deletion"
flag set.

While analyzing the a recognition result (in simonmodeltest) these
leaves were inserted at the end of the list of result nodes of the test
sample while processing it. This caused the algorithm to count them
again - this time as insertion errors - when it encountered the new,
trailing nodes.

Example:
Prompt: Foo Bar
Result: Bar
Algorithm:
0. Nodes: {Bar}
1. i = 0; Foo is missing, Bar is okay -> Nodes: {Bar[Okay], Foo[Deletion]}
2. i = 1; Foo at the end is too much -> Nodes {Bar[Okay], Foo[Insertion,
Deletion]}
2. i = 2; Done

After the fix:
0. Nodes: {Bar}
1. i = 0; Foo is missing, Bar is okay -> Nodes: {Foo[Deletion], Bar[Okay]}
          i = 2
2. i = 2; Done
BUG: 310099