Avoid crash when "deleting" non existant scenarios
The GHNS dialog reports things that were installed but removed
immediately (before ever accepting the GHNS dialog) as both installed
and removed.
That means that Simon would try to add items for this non-existant
scenario to the list of available scenarios but fail (silently). Then it
would try to remove the scenario for which there is no item which caused
a crash.
BUG: 313417