Implementing Carbon Backend for Event Simulation on Mac OS X
Some responsibilities were moved from the implementation to the
CoreEvents base class (like keeping track of keyboard
modifier's state). A newly introduced PCEvents groups functionality
that Linux and Windows share, but that does not apply to OS X
(right now this is only the handling of "dead keys", which work
significantly differently on OS X).